+++ src/index.js
@@ -1,6 +1,7 @@
 import find from './find/index.js'
 import display from './display/index.js'
 import version from './generated/version.js'
+import { canonicalize, zones } from './data/index.js'
 
 const soft = function (str) {
   if (typeof str !== 'string') {
@@ -10,7 +11,8 @@ const soft = function (str) {
   if (typeof ids === 'string') {
     ids = [ids]
   }
-  ids = ids.map((id) => display(id, str))
+  ids = [...new Set(ids.map(canonicalize))].filter(id => Object.hasOwn(zones, id))
+  ids = ids.map(id => display(id))
   return ids
 }
 soft.version = version